Fixed position elements

The affix utility function applies Bootstrap classes to fix elements to the top or bottom of a page. Use "sticky" to cause an element to fix to the top of a page after the element is scrolled past. Important, the IE11 and Edge browsers do not support the sticky behavior.

affix(x, position)

Arguments

x
A tag element or .style pronoun.

position
One of "top", "bottom", or "sticky" specifying the fixed behavior of an element.

See also