alert {yonder}	R Documentation

Alert boxes

Description

Use an alert element to let the user know of successes or to call attention to problems.

Usage


alert(..., dismissible = TRUE, fade = TRUE)

Arguments

...
Character strings specifying the text of the alert or additional named arguments passed as HTML attributes to the alert element.

dismissible
One of TRUE or FALSE specifying if the alert may be dismissed by the user, defaults to TRUE.

fade
One of TRUE or FALSE specifying if the alert fades out or immediately disappears when dismissed, defaults to TRUE.

See Also

Other components: badge(), blockquote(), card(), collapsePane(), d1(), dropdown(), img(), jumbotron(), modal(), navContent(), popover(), pre(), toast()

Examples



### Default alert

alert("Donec at pede.") %>%
  background("blue")

### A more complex alert

alert(
  h4("Etiam vel tortor sodales"),
  hr(),
  p("Fusce commodo.")
) %>%
  background("amber")


[Package yonder version 0.2.0 Index]