Button group inputs

A set of buttons with custom values.

buttonGroupInput(
  ...,
  id,
  choices = NULL,
  values = choices,
  labels = deprecated(),
  direction = "horizontal"
)

updateButtonGroupInput(
  id,
  choices = NULL,
  values = choices,
  enable = NULL,
  disable = NULL,
  session = getDefaultReactiveDomain()
)

Arguments

...
Additional named arguments passed as HTML attributes to the parent element.

id
A character string specifying the id of the reactive input.

choices
A character vector specifying the labels for each button in the group.

values
A character vector specifying the values of each button in the group, defaults to choices.

labels
Deprecated, see values.

direction
One of "horizontal" or "vertical" specifying the layout of the buttons, defaults to "horizontal".

enable
One of values indicating individual buttons to enable or TRUE to enable the entire input, defaults to NULL.

disable
One of values indicating individual buttons to disable or TRUE to disable the entire input, defaults to NULL.

session
A reactive context, defaults to getDefaultReactiveDomain().

See also

Examples

### Default input

buttonGroupInput(
  id = "group1",
  choices = c("Once", "Twice", "Thrice"),
  values = c(1, 2, 3)
)

#> <div class="yonder-button-group btn-group" id="group1" role="group">
#>   <button type="button" class="btn" value="1">Once</button>
#>   <button type="button" class="btn" value="2">Twice</button>
#>   <button type="button" class="btn" value="3">Thrice</button>
#> </div>

### Styling the button group

buttonGroupInput(
  .style %>%
    background("primary"),
  id = "group2",
  choices = c("Button 1", "Button 2", "Button 3")
)

#> <div class="yonder-button-group btn-group btn-group-primary" id="group2" role="group">
#>   <button type="button" class="btn" value="Button 1">Button 1</button>
#>   <button type="button" class="btn" value="Button 2">Button 2</button>
#>   <button type="button" class="btn" value="Button 3">Button 3</button>
#> </div>