d1 {yonder}	R Documentation

Headings

Description

Display headings are not meant to replace the standard HTML heading tags, they are a stand out alternative for eye-catching titles.

Usage


d1(...)

d2(...)

d3(...)

d4(...)

Arguments

...
Any number of character strings or tag elements or named arguments passed as HTML attributes to the parent element.

See Also

Other components: alert(), badge(), blockquote(), card(), collapsePane(), dropdown(), img(), jumbotron(), modal(), navContent(), popover(), pre(), toast()

Examples



### d1

d1("Eye-catching!")

### d2

d2("Just incredible")

### d3

d3("Wowie, zowie")

### d4

d4("You'll never guess what happens next.")


[Package yonder version 0.2.0 Index]