Display

Use the display() utility to adjust how a tag element is rendered. All arguments are responsive allowing you to hide elements on small screens or convert elements from inline to block on large screens.

display(x, type)

Arguments

x
A tag element or .style pronoun.

type
A responsive argument. One of "inline", "block", "inline-block", "flex", "inline-flex", or "none".

See also