Floats

Use float() to float an element to the left or right side of its parent element. A newspaper layout is a classic example where an image is floated with text wrapped around.

float(x, side)

Arguments

x
A tag element or .style pronoun.

side
A responsive argument. One of "left" or "right" specifying the side to float the element.

See also

Examples

### Float an alert

div(
  .style %>%
    padding(2),
  alert("This just in!") %>%
    margin(3) %>%
    background("warning") %>%
    float("left"),
  p("Fusce commodo. Nullam tempus. Nunc rutrum turpis sed pede.",
    "Phasellus lacus.  Cras placerat accumsan nulla.",
    "Fusce sagittis, libero non molestie mollis, ",
    "magna orci ultrices dolor, at vulputate neque nulla lacinia eros."),
  p("Nulla facilisis, risus a rhoncus fermentum, tellus tellus",
    "lacinia purus, et dictum nunc justo sit amet elit."),
  p("Proin neque massa, cursus ut, gravida ut, lobortis eget, lacus.",
    "Aliquam posuere.",
    "Sed id ligula quis est convallis tempor."),
  p("Fusce dapibus, tellus ac cursus commodo, ",
    "tortor mauris condimentum nibh, ut fermentum massa justo sit",
    "amet risus.")
)

#> <div class="p-2">
#>   <div class="alert alert-dismissible fade show m-3 alert-warning float-left" role="alert">
#>     This just in!
#>     <button type="button" class="close" data-dismiss="alert" aria-label="Close">
#>       <span aria-hidden="true">&times;</span>
#>     </button>
#>   </div>
#>   <p>
#>     Fusce commodo. Nullam tempus. Nunc rutrum turpis sed pede.
#>     Phasellus lacus.  Cras placerat accumsan nulla.
#>     Fusce sagittis, libero non molestie mollis, 
#>     magna orci ultrices dolor, at vulputate neque nulla lacinia eros.
#>   </p>
#>   <p>
#>     Nulla facilisis, risus a rhoncus fermentum, tellus tellus
#>     lacinia purus, et dictum nunc justo sit amet elit.
#>   </p>
#>   <p>
#>     Proin neque massa, cursus ut, gravida ut, lobortis eget, lacus.
#>     Aliquam posuere.
#>     Sed id ligula quis est convallis tempor.
#>   </p>
#>   <p>
#>     Fusce dapibus, tellus ac cursus commodo, 
#>     tortor mauris condimentum nibh, ut fermentum massa justo sit
#>     amet risus.
#>   </p>
#> </div>