Font color, size, weight

The font() utility modifies the color, size, weight, case, or alignment of a tag element's text. All arguments default to NULL, in which case they are ignored. For example, font(.., size = "lg") increases font size without affecting color, weight, case, or alignment.

font(x, color = NULL, size = NULL, weight = NULL, case = NULL, align = NULL)

Arguments

x
A tag element or .style pronoun.

color
One of "primary", "secondary", "success", "info", "warning", "danger", "light" or "dark" specifying the text color of the tag element, defaults to NULL.

size
Deprecated, in future versions of bootstrap resonsive font sizing will be enabled by default, https://github.com/twbs/bootstrap/pull/29152.

One of "xs", "sm", "base", "lg", "xl" specifying a font size relative to the default base page font size, defaults to NULL.

weight
One of "bold", "bolder", "normal", "lighter", or "light" specifying the font weight of the element's text, defaults to NULL. "bolder" and "lighter" change the weight relative to the current font weight of the element.

case
One of "upper", "lower", or "title" specifying a transformation of the tag element's text, default to NULL.

align
A responsive argument. One of "left", "center", "right", or "justify", specifying the alignment of the tag element's text, defaults to NULL.

See also

Examples

### Changing text color

card(
  .style %>%
    font("primary") %>%
    border("primary"),
  header = h3("Important!"),
  p("This is a reminder.")
)

#> <div class="card text-primary border border-primary">
#>   <div class="card-header">
#>     <h3>Important!</h3>
#>   </div>
#>   <div class="card-body">
#>     <p class="card-text">This is a reminder.</p>
#>   </div>
#> </div>

### Changing font weight

p("Curabitur lacinia pulvinar nibh.") %>%
  font(weight = "bold")

#> <p class="font-weight-bold">Curabitur lacinia pulvinar nibh.</p>

p("Proin quam nisl, tincidunt et.") %>%
  font(weight = "light")

#> <p class="font-weight-light">Proin quam nisl, tincidunt et.</p>