Responsive images and figures

A small update to tags$img and tags$figure. Create responsive images with img. figure has specific arguments for an image child element and image caption.

img(src, ...)

figure(image, caption = NULL, ...)

Arguments

src
A character string specifying the source of the image.

...
Additional tag elements or named arguments passed as HTML attributes to the parent element.

image
An <img> tag, typically a call to img.

caption
A character string specifying the image caption, defaults to NULL.

See also