jumbotron {yonder}	R Documentation

Jumbotron

Description

A showcase banner, good for front or splash pages.

Usage


jumbotron(..., title = NULL, subtitle = NULL)

Arguments

...
Tag elements passed as child elements or named arguments passed as HTML attributes to the parent element.

title
A character string specifying a title for the jumbotron, defaults to NULL, in which case a title is not added.

subtitle
A character string specifying a subtitle for the jumbotron, defaults to NULL, in which case a subtitle is not added.

See Also

Other components: alert(), badge(), blockquote(), card(), collapsePane(), d1(), dropdown(), img(), modal(), navContent(), popover(), pre(), toast()

Examples



### Landing page welcome

jumbotron(
  title = "Welcome, welcome!",
  subtitle = "Here we are showcasing the very showcase itself.",
  tags$p(
    "Now let's talk more about that superb new feature."
  )
)


[Package yonder version 0.2.0 Index]