menuInput {yonder}	R Documentation

Menu inputs

Description

A toggleable dropdown menu input. Menu inputs may be used as standalone reactive inputs or within a navInput(). For building custom, more complex dropdown elements please see dropdown().

Usage


menuInput(
  id,
  label,
  choices = NULL,
  values = choices,
  selected = NULL,
  ...,
  direction = "down",
  align = "left"
)

updateMenuInput(
  id,
  label = NULL,
  choices = NULL,
  values = choices,
  selected = NULL,
  enable = NULL,
  disable = NULL,
  session = getDefaultReactiveDomain()
)

Arguments

id
A character string specifying the id of the reactive input.

label
A character string or tag element specifying the label of the menu's toggle button.

choices
A character vector specifying the choice text of the menu's items.

values
A character vector specifying the values of the menu's items, defaults to choices.

selected
One or more of values specifying which choices are selected by default, defaults to NULL, in which case no choices are initially selected.

...
Additional named arguments passed as HTML attributes to the parent element or tag elements passed as child elements to the parent element.

direction
One of "up", "right", "down", or "left" specifying which direction the menu opens, defaults to "down".

align
One or "right" or "left" specifying which side of the toggle button the menu aligns to, defaults to "left".'

enable
One of values specifying particular choices to enable or TRUE specifying the entire input is enabled, defaults to NULL.

disable
One of values specifying particular choices to disable or TRUE specifying the entire input is disabled, defaults to NULL.

session
A reactive context, defaults to getDefaultReactiveDomain().

See Also

Other inputs: buttonGroupInput(), buttonInput(), checkbarInput(), checkboxInput(), chipInput(), fileInput(), formInput(), listGroupInput(), navInput(), radioInput(), radiobarInput(), rangeInput(), selectInput(), textInput()

Examples



### A simple menu

menuInput(
  id = "menu1",
  label = "Menu",
  choices = c(
    "Choice 1",
    "Choice 2",
    "Choice 3"
  )
)

### Use in navigation

navInput(
  id = "nav1",
  choices = list(
    "Tab 1",
    menuInput(
      id = "navOptions",
      label = "Tab 2",
      choices = c(
        "Option 1",
        "Option 2",
        "Option 3"
      )
    ),
    "Tab 3",
    "Tab 4"
  ),
  values = paste0("tab", 1:4)
)


[Package yonder version 0.2.0 Index]