Scrollable code snippets

The pre function adds a maximum height and scroll bar to the standard <pre> element.

pre(...)

Arguments

...
Text, tag elements, or named arguments passed as HTML attributes to the tag.

See also

Examples

### Simple example

pre(
  "shinyApp(",
  "  ui = container(",
  "    columns(",
  "      column(",
  "      ",
  "      )",
  "    )",
  "  )",
  "  server = function(input, output) {",
  "  ",
  "  }",
  ")"
)

#> <pre class="pre-scrollable">
#>   shinyApp(
#>     ui = container(
#>       columns(
#>         column(
#>         
#>         )
#>       )
#>     )
#>     server = function(input, output) {
#>     
#>     }
#>   )
#> </pre>