Radiobar inputs

A stylized group of radio inputs. A radiobar input is similar to a button group, but with a checked or highlighted stated. Additionally, only one value at most may be selected at any given time.

radiobarInput(id, choices, values = choices, selected = values[[1]], ...)

updateRadiobarInput(
  id,
  choices = NULL,
  values = choices,
  selected = NULL,
  enable = NULL,
  disable = NULL,
  session = getDefaultReactiveDomain()
)

Arguments

id
A character string specifying the id of the reactive input.

choices
A character vector or list of tag elements specifying the labels of the input's choices.

values
A vector specifying the values of the input's choices, defaults to choices.

selected
One of values specifying the input's default selected choice, defualts to values[[1]].

...
Additional named arguments passed as HTML attributes to the parent element or tag elements passed as child elements to the parent element.

enable
One of values specifying particular choices to enable or TRUE specifying the entire input is enabled, defaults to NULL.

disable
One of values specifying particular choices to disable or TRUE specifying the entire input is disabled, defaults to NULL.

session
A reactive context, defaults to getDefaultReactiveDomain().

See also

Examples

### Radiobars

radiobarInput(
  id = "radiobar1",
  choices = c(
    "fusce sagittis",
    "libero non molestie",
    "magna orci",
    "ultrices dolor"
  ),
  selected = "ultrices dolor"
) %>%
  background("secondary")

#> <div class="yonder-radiobar btn-group btn-group-toggle d-flex bg-secondary" id="radiobar1" data-toggle="buttons">
#>   <label class="btn">
#>     <input name="radiobar1" type="radio" value="fusce sagittis" autocomplete="off"/>
#>     fusce sagittis
#>   </label>
#>   <label class="btn">
#>     <input name="radiobar1" type="radio" value="libero non molestie" autocomplete="off"/>
#>     libero non molestie
#>   </label>
#>   <label class="btn">
#>     <input name="radiobar1" type="radio" value="magna orci" autocomplete="off"/>
#>     magna orci
#>   </label>
#>   <label class="btn active">
#>     <input name="radiobar1" type="radio" value="ultrices dolor" checked autocomplete="off"/>
#>     ultrices dolor
#>   </label>
#> </div>