replaceContent {yonder}	R Documentation

Dynamic content

Description

An application may require dynamic content. This content may be quite simple. The content could also be quite variable. These tools are an alternative to the standard output related render*() functions.

Usage


replaceContent(id, ..., session = getDefaultReactiveDomain())

removeContent(id, session = getDefaultReactiveDomain())

Arguments

id
A character string specifying a reactive id.

...
Additional named arguments passed as HTML attributes to the parent element or unnamed arguments passed as the new contents of the output element.

session
A reactive context, defaults to getDefaultReactiveDomain().

Details

These functions are experimental and are subject to change. Additionally, they may be moved from this package entirely.

[Package yonder version 0.2.0 Index]