Textual inputs

Different types of textual inputs are provided to best support mobile keyboards and assistive technologies. A password input will mask its contents. Email inputs offer client-side validation depending on the browser.

textInput(id, value = NULL, placeholder = NULL, ...)

updateTextInput(id, value = NULL, enable = NULL, disable = NULL,
  valid = NULL, invalid = NULL, session = getDefaultReactiveDomain())

groupTextInput(id, value = NULL, placeholder = NULL, ...,
  left = NULL, right = NULL)

updateGroupTextInput(id, value = NULL, enable = NULL, disable = NULL,
  valid = NULL, invalid = NULL, session = getDefaultReactiveDomain())

searchInput(id, value = NULL, placeholder = NULL, ...)

updateSearchInput(id, value = NULL, enable = NULL, disable = NULL,
  valid = NULL, invalid = NULL, session = getDefaultReactiveDomain())

emailInput(id, value = NULL, placeholder = NULL, ...)

updateEmailInput(id, value = NULL, enable = NULL, disable = NULL,
  valid = NULL, invalid = NULL, session = getDefaultReactiveDomain())

urlInput(id, value = NULL, placeholder = NULL, ...)

updateUrlInput(id, value = NULL, enable = NULL, disable = NULL,
  valid = NULL, invalid = NULL, session = getDefaultReactiveDomain())

telephoneInput(id, value = NULL, placeholder = NULL, ...)

updateTelephoneInput(id, value = NULL, enable = NULL, disable = NULL,
  valid = NULL, invalid = NULL, session = getDefaultReactiveDomain())

passwordInput(id, value = NULL, placeholder = NULL, ...)

updatePassowrdInput(id, value = NULL, enable = NULL, disable = NULL,
  valid = NULL, invalid = NULL, session = getDefaultReactiveDomain())

numberInput(id, value = NULL, placeholder = NULL, ...)

updateNumberInput(id, value = NULL, enable = NULL, disable = NULL,
  valid = NULL, invalid = NULL, session = getDefaultReactiveDomain())

Arguments

id
A character string specifying the id of the reactive input.

value
A character string or a value coerced to a character string specifying the default value of the textual input.

placeholder
A character string specifying placeholder text for the input, defaults to NULL, in which case there is no placeholder text.

...
Additional named arguments passed as HTML attributes to the parent element or tag elements passed as child elements to the parent element.

enable
One of values specifying particular choices to enable or TRUE specifying the entire input is enabled, defaults to NULL.

disable
One of values specifying particular choices to disable or TRUE specifying the entire input is disabled, defaults to NULL.

valid
A character string specifying a message to the user indicating how the input's value is valid, defaults to NULL.

invalid
A character string specifying a message to the user indicating how the input's value is invalid, defaults to NULL.

session
A reactive context, defaults to getDefaultReactiveDomain().

left
A character vector specifying static addons or buttonInput() or dropdown() elements specifying dynamic addons. Addons affect the reactive value of the group input, see the Details section below for more information.

left is character or right is character

If left or right are character vectors, then the group input functions like a text input. The value will update and trigger a reactive event when the text box is modified. The group input's reactive value is the concatention of the static addons specified by left or right and the value of the text input.

left is button or right is button

The button does not change the value of the group input. However, the input no longer triggers event when the text box is updated. Instead the value is updated when a button is clicked. Static addons are still applied to the group input value.

left is a dropdown or right is a dropdown

The value of the group input does chance depending on the clicked dropdown menu item. The value of the input group is the concatentation of the dropdown input value, the value of the text input, and any static addons.

right
A character vector specifying static addons or buttonInput() or dropdown() elements specifying dynamic addons. Addons affect the reactive value of the group input, see the Details section below for more information.

left is character or right is character

If left or right are character vectors, then the group input functions like a text input. The value will update and trigger a reactive event when the text box is modified. The group input's reactive value is the concatention of the static addons specified by left or right and the value of the text input.

left is button or right is button

The button does not change the value of the group input. However, the input no longer triggers event when the text box is updated. Instead the value is updated when a button is clicked. Static addons are still applied to the group input value.

left is a dropdown or right is a dropdown

The value of the group input does chance depending on the clicked dropdown menu item. The value of the input group is the concatentation of the dropdown input value, the value of the text input, and any static addons.

See also

Examples

### Basic text

textInput(id = "text")

#> <div class="yonder-textual" id="text">
#>   <input class="form-control" type="text" autocomplete="off"/>
#>   <div class="valid-feedback"></div>
#>   <div class="invalid-feedback"></div>
#> </div>

### Search

searchInput(id = "search")

#> <div class="yonder-textual" id="search">
#>   <input class="form-control" type="search" autocomplete="off"/>
#>   <div class="valid-feedback"></div>
#>   <div class="invalid-feedback"></div>
#> </div>

### Email

emailInput(id = "email")

#> <div class="yonder-textual" id="email">
#>   <input class="form-control" type="email" autocomplete="off"/>
#>   <div class="valid-feedback"></div>
#>   <div class="invalid-feedback"></div>
#> </div>

### URLs

urlInput(id = "url")

#> <div class="yonder-textual" id="url">
#>   <input class="form-control" type="url" autocomplete="off"/>
#>   <div class="valid-feedback"></div>
#>   <div class="invalid-feedback"></div>
#> </div>

### Telephone numbers

telephoneInput(id = "tele")

#> <div class="yonder-textual" id="tele">
#>   <input class="form-control" type="tel" autocomplete="off"/>
#>   <div class="valid-feedback"></div>
#>   <div class="invalid-feedback"></div>
#> </div>

### Passwords

passwordInput(id = "password")

#> <div class="yonder-textual" id="password">
#>   <input class="form-control" type="password" autocomplete="off"/>
#>   <div class="valid-feedback"></div>
#>   <div class="invalid-feedback"></div>
#> </div>

### Numbers

numberInput(id = "num")

#> <div class="yonder-textual" id="num">
#>   <input class="form-control" type="number" autocomplete="off"/>
#>   <div class="valid-feedback"></div>
#>   <div class="invalid-feedback"></div>
#> </div>