User interface

Begin creating a user interface. The webpage() function properly lays out a navigation bar and main section of elements.

webpage(..., nav = NULL)

Arguments

...
Any number of tag elements or named values added as children and attributes to the main section of the page.

nav
A navigation element, typically a call to navbar(), added at the top of the page, defaults to NULL.

See also

Examples

webpage(
  p("Pretty simple")
)

#> <body>
#>   <header></header>
#>   <main role="main">
#>     <p>Pretty simple</p>
#>   </main>
#> </body>

webpage(
  nav = navbar(),
  container(
    columns(
      column(),
      column()
    )
  )
)

#> <body>
#>   <header>
#>     <nav class="navbar navbar-expand navbar-light">
#>       <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#nav-content-265-150" aria-controls="nav-content-265-150" aria-expanded="false" aria-label="Toggle navigation">
#>         <span class="navbar-toggler-icon"></span>
#>       </button>
#>       <div class="collapse navbar-collapse" id="nav-content-265-150"></div>
#>     </nav>
#>   </header>
#>   <main role="main">
#>     <div class="container-fluid">
#>       <div class="row">
#>         <div class="col"></div>
#>         <div class="col"></div>
#>       </div>
#>     </div>
#>   </main>
#> </body>