blockquote {yonder}	R Documentation

Blockquotes

Description

Stylized blockquotes, an updated builder function for <blockquote>.

Usage


blockquote(..., source = NULL, align = "left")

Arguments

...
Any number of tags elements or character strings passed as child elements or named arguments passed as HTML attributes to the parent element.

source
The quote source, use tags$cite to format the source title, defaults to NULL.

align
One of "left" or "right", defaults to "left".

See Also

Other components: alert(), badge(), card(), collapsePane(), d1(), dropdown(), img(), jumbotron(), modal(), navContent(), popover(), pre(), toast()

Examples



### Simple example

blockquote(
  "Anyone can love a thing because.",
  "That's as easy as putting a penny in your pocket.",
  "But to love something despite.",
  "To know the flaws and love them too.",
  "That is rare and pure and perfect.",
  source = list(
    "Patrick Rothfuss,", tags$cite("The Wise Man's Fear")
  )
)


[Package yonder version 0.2.0 Index]