Button and link inputs

Button inputs are useful as triggers for reactive or observer expressions. The reactive value of a button input begins as NULL, but subsequently is the number of clicks.

buttonInput(..., id, label, stretch = FALSE, download = FALSE, tooltip = NULL)

updateButtonInput(
  id,
  label = NULL,
  value = NULL,
  disable = NULL,
  enable = NULL,
  tooltip = NULL,
  session = getDefaultReactiveDomain()
)

linkInput(..., id, label, stretch = FALSE, download = FALSE, tooltip = NULL)

updateLinkInput(
  id,
  label = NULL,
  value = NULL,
  enable = NULL,
  disable = NULL,
  tooltip = NULL,
  session = getDefaultReactiveDomain()
)

tooltip(..., placement = "top", fade = TRUE)

Arguments

...
Additional named arguments passed as HTML attributes to the parent element.

id
A character string specifying the id of the reactive input.

label
A character string specifying the label text on the button or link input.

stretch
One of TRUE or FALSE specifying stretched behaviour for the button or link input, defaults to FALSE. If TRUE, the button or link will receive clicks from its containing block element. For example, a stretched button or link inside a card() would update whenever the user clicked on the card.

download
One of TRUE or FALSE specifying if the button or link input is used to trigger a download, defaults to FALSE.

tooltip
A call to tooltip() specifying a tooltip for the button or link input, defaults to NULL.

value
A number specifying a new value for the button, defaults to NULL.

disable
if TRUE the button is disabled and will not react to clicks from the user, default to NULL.

enable
If TRUE the button is enabled and will react to clicks from the user, defaults to NULL.

session
A reactive context, defaults to getDefaultReactiveDomain().

placement
One of "top", "right", "bottom", or "left" specifying what side of the tag element the tooltip appears on.

fade
One of TRUE or FALSE specifying if the tooltip fades in when shown and fades out when hidden, defaults to TRUE.

Details

Tooltips

To remove a button or link input's tooltip pass an empty tooltip, tooltip(), to updateButtonInput() or updateLinkInput().

See also

Examples

### A plain button

buttonInput(
  id = "button1",
  label = "Simple"
)

#> <button class="yonder-button btn" type="button" role="button" id="button1" autocomplete="off">Simple</button>

### Add a little color

buttonInput(
  .style %>%
    background("primary"),
  id = "button2",
  label = "Blue"
)

#> <button autocomplete="off" class="yonder-button btn btn-primary" id="button2" role="button" type="button">Blue</button>

# Alternatively, a button can fill the width of its parent element.

buttonInput(
  .style %>%
    background("danger"),
  id = "button3",
  label = "Full-width",
  fill = TRUE  # <-
)

#> <button autocomplete="off" class="yonder-button btn btn-danger" fill="TRUE" id="button3" role="button" type="button">Full-width</button>

# Use design utilities to further adjust the width of a button.

buttonInput(
  .style %>%
    background("danger") %>%
    width(75),  # <-
  id = "button4",
  label = "Full and back again",
  fill = TRUE  # <-
)

#> <button autocomplete="off" class="yonder-button btn btn-danger w-75" fill="TRUE" id="button4" role="button" type="button">Full and back again</button>

### Reactive links

div("Curabitur", linkInput(id = "link1", label = "vulputate"),
    "vestibulum lorem.")

#> <div>
#>   Curabitur
#>   <button class="yonder-link btn btn-link" id="link1">vulputate</button>
#>   vestibulum lorem.
#> </div>

### Stretched buttons and links

card(
  header = "Card with stretched button",
  p("Notice when you hover over the card, the button also detects ",
    "the hover."),
  buttonInput(
    .style %>%
      background("primary"),
    id = "go",
    label = "Go go go",
    stretch = TRUE
  )
)

#> <div class="card">
#>   <div class="card-header">Card with stretched button</div>
#>   <div class="card-body">
#>     <p class="card-text">
#>       Notice when you hover over the card, the button also detects 
#>       the hover.
#>     </p>
#>   </div>
#>   <div class="card-body">
#>     <button autocomplete="off" class="yonder-button btn stretched-link btn-primary" id="go" role="button" type="button">Go go go</button>
#>   </div>
#> </div>

### Download button

buttonInput(
  download = TRUE,
  id = "download1",
  label = "Download",
  icon("download")
)

#> <a class="yonder-button btn shiny-download-link" type="button" role="button" href="" _target download id="download1" autocomplete="off">
#>   Download
#>   <i class="fa fa-download"></i>
#> </a>