img {yonder}	R Documentation

Responsive images and figures

Description

A small update to tags$img and tags$figure. Create responsive images with img. figure has specific arguments for an image child element and image caption.

Usage


img(src, ...)

figure(image, caption = NULL, ...)

Arguments

src
A character string specifying the source of the image.

...
Additional tag elements or named arguments passed as HTML attributes to the parent element.

image
An <img> tag, typically a call to img.

caption
A character string specifying the image caption, defaults to NULL.

See Also

Other components: alert(), badge(), blockquote(), card(), collapsePane(), d1(), dropdown(), jumbotron(), modal(), navContent(), popover(), pre(), toast()

[Package yonder version 0.2.0 Index]