navbar {yonder}	R Documentation

Navigation bar

Description

Add a navigation bar to your application with navbar(). Navigation bars may include a tab toggle (useful for multi-page applications), inline forms (perhaps a search feature or login item), or character strings to add simple text. Navbars are responsive and will collapse on small screens, think mobile device. A button is automatically added to toggle between the collapsed and expanded states.

Usage


navbar(..., brand = NULL, collapse = NULL)

Arguments

...
A tab toggle, inline forms, or text to add to include as part of the navigation bar.

brand
A tag element or text placed on the left end of the navbar, defaults to NULL, in which case nothing is added.

collapse
One of "sm", "md", "lg", "xl", or NULL specifying the breakpoint at which the navbar collaspes, defaults to NULL, in which case the navbar is always expanded.

See Also

Other layout functions: column(), fieldset(), flex(), responsive, webpage()

Examples



### Navbar with tabs

div(
  navbar(
    brand = "Navbar",
    navInput(
      id = "tabs",
      choices = c("Home", "About", "Our process")
    ) %>%
      margin(right = "auto"),
    formInput(
      inline = TRUE,
      id = "navForm",
      textInput(
        type = "search",
        id = "site_search",
        placeholder = "Search"
      ) %>%
        margin(right = c(sm = 2)),
      formSubmit(
        label = "Search",
        value = "search"
      ) %>%
        background("amber")
    )
  ) %>%
    background("teal"),
  container(
    navContent(
      navPane(
        h3("Home")
      ),
      navPane(
        h3("About")
      ),
      navPane(
        h3("The process")
      )
    )
  )
)


[Package yonder version 0.2.0 Index]