pre {yonder}	R Documentation

Scrollable code snippets

Description

The pre function adds a maximum height and scroll bar to the standard <pre> element.

Usage


pre(...)

Arguments

...
Text, tag elements, or named arguments passed as HTML attributes to the tag.

See Also

Other components: alert(), badge(), blockquote(), card(), collapsePane(), d1(), dropdown(), img(), jumbotron(), modal(), navContent(), popover(), toast()

Examples



### Simple example

pre(
  "shinyApp(",
  "  ui = container(",
  "    columns(",
  "      column(",
  "      ",
  "      )",
  "    )",
  "  )",
  "  server = function(input, output) {",
  "  ",
  "  }",
  ")"
)


[Package yonder version 0.2.0 Index]