rangeInput {yonder}	R Documentation

Range input

Description

rangeInput() creates a simple numeric range input.

Usage


rangeInput(id, min = 0, max = 100, default = min, step = 1, ...)

updateRangeInput(
  id,
  value = NULL,
  enable = NULL,
  disable = NULL,
  session = getDefaultReactiveDomain()
)

Arguments

id
A character string specifying the id of the reactive input.

min
A number specifying the minimum value of the input, defaults to 0.

max
A number specifying the maximum value of the input, defaults to 100.

default
A number between min and max specifying the default value of the input, defaults to min.

step
A number specifying the interval step of the input, defaults to 1.

...
Additional named arguments passed as HTML attributes to the parent element or tag elements passed as child elements to the parent element.

value
A number specifying a new value for the input, defaults to NULL.

enable
One of values specifying particular choices to enable or TRUE specifying the entire input is enabled, defaults to NULL.

disable
One of values specifying particular choices to disable or TRUE specifying the entire input is disabled, defaults to NULL.

session
A reactive context, defaults to getDefaultReactiveDomain().

Details

The sophistication of this input will improve as browsers adopt the latest HTML standards.

See Also

Other inputs: buttonGroupInput(), buttonInput(), checkbarInput(), checkboxInput(), chipInput(), fileInput(), formInput(), listGroupInput(), menuInput(), navInput(), radioInput(), radiobarInput(), selectInput(), textInput()

Examples



### Range inputs

# Select from a range of numeric values.

rangeInput(id = "range1")

### Custom step

rangeInput(id = "range2", step = 10)


[Package yonder version 0.2.0 Index]