webpage {yonder}	R Documentation

User interface

Description

Begin creating a user interface. The webpage() function properly lays out a navigation bar and main section of elements.

Usage


webpage(..., nav = NULL)

Arguments

...
Any number of tag elements or named values added as children and attributes to the main section of the page.

nav
A navigation element, typically a call to navbar(), added at the top of the page, defaults to NULL.

See Also

Other layout functions: column(), fieldset(), flex(), navbar(), responsive

Examples



webpage(
  p("Pretty simple")
)

webpage(
  nav = navbar(),
  container(
    columns(
      column(),
      column()
    )
  )
)


[Package yonder version 0.2.0 Index]