yonder-package {yonder}	R Documentation

An opinionated shiny framework

Description

Yonder is a set of tools for flexible and creative shiny application construction and design.

Inputs

Yonder provides many familiar inputs like selectInput() or radioInput(). There are also new inputs like groupTextInput() or formInput().

Changes to be mindful of

Input functions have an id argument instead of inputId.

Input functions do not include a label argument for the purpose of adding a label above the input. Button and menu inputs do include a label argument, but these arguments refer to button labels. If you would like to add a label above an input please use formGroup().

Familiar variants

Looking for ... ?

actionButton() or actionLink() use buttonInput() or linkInput()

radioButtons() use radioInput()

checkboxGroupInput() use checkbarInput() or checkboxInput()

numericInput() use numberInput()

selectizeInput() use selectInput() or chipInput()

submitButton() use formInput() and formSubmit()

Layout and content

Included are a handful of utilities for building applications suited for devices and screens of varying sizes. For real control over spacing elements be sure to check out flex(), which gives you the power of flexbox layout.

Familiar variants

Looking for ... ?

fluidRow() or fixedRow() use columns()

fixedPage(), fluidPage(), or sidebarLayout() use container(), columns(), and column()

navbarPage() use navbar()

tabPanel() use navContent() and navPane()

modalDialog() use modal()

Author(s)

Maintainer: Nathan Teetor nathanteetor@gmail.com

Other contributors:

The Bootstrap Authors (Bootstrap library) [copyright holder]

Twitter, Inc (Bootstrap library) [copyright holder]

JS Foundation (jQuery library) [copyright holder]

Federico Zivolo (popper.js library) [contributor, copyright holder]

Johann Servoire (bs-custom-file-input library) [contributor, copyright holder]

See Also

Useful links:

https://nteetor.github.io/yonder

Report bugs at https://github.com/nteetor/yonder/issues

[Package yonder version 0.2.0 Index]